Rename `rpm-package-function` to `azure-blobrepo-rpm`
This project has been renamed to better reflect its purpose and the Azure services it uses.
