+++ azure_blobrepo_rpm/tooling/azcmd.py
